move GeneTEA-related API functions to legacyPortalAPI (#334)
