+++ frontend/packages/@depmap/api/src/legacyPortalAPI/index.ts
@@ -6,6 +6,7 @@ import * as context_explorer from "./resources/context_explorer";
 import * as data_page from "./resources/data_page";
 import * as download from "./resources/download";
 import * as entity_summary from "./resources/entity_summary";
+import * as genetea from "./resources/genetea";
 import * as interactive from "./resources/interactive";
 import * as tda from "./resources/tda";
 import * as misc from "./resources/misc";
@@ -19,6 +20,7 @@ export const legacyPortalAPI = {
   ...data_page,
   ...download,
   ...entity_summary,
+  ...genetea,
   ...interactive,
   ...tda,
   ...misc,
