+++ frontend/packages/portal-frontend/src/data-explorer-2/deprecated-api.ts
@@ -278,103 +278,6 @@ export function fetchDimensionLabelsToDatasetsMapping(
   return fetchJson(`/dimension_labels_to_datasets_mapping?${query}`);
 }
 
-export async function fetchGeneTeaEnrichment(
-  genes: string[],
-  limit: number | null
-): Promise<{
-  term: string[];
-  synonyms: string[][];
-  coincident: string[][];
-  fdr: number[];
-  matchingGenes: string[][];
-  total: number;
-}> {
-  const query = qs.stringify(
-    {
-      gene_list: genes,
-      remove_overlapping: "true",
-      n: limit || -1,
-      model: "v2",
-    },
-    { arrayFormat: "repeat" }
-  );
-
-  interface RawResponse {
-    // TODO: Give the user feedback when some genes are invalid.
-    invalid_genes: string[];
-    total_n_enriched_terms: number;
-    enriched_terms: {
-      Term: string[];
-      // semicolon separated strings
-      Synonyms: (string | null)[];
-      // semicolon separated strings
-      "Coincident Terms": (string | null)[];
-      FDR: number[];
-      // Gene lists are just space-separated strings like "ADSL CAD UMPS"
-      "Matching Genes in List": string[];
-    };
-  }
-
-  const transformResponse = (body: RawResponse) => {
-    // `enriched_terms` can be null when there are no relevant terms. We'll
-    // return a wrapper object to distinguish this from some kind of error.
-    if (body.enriched_terms === null) {
-      return {
-        term: [],
-        synonyms: [],
-        coincident: [],
-        fdr: [],
-        matchingGenes: [],
-        total: 0,
-      };
-    }
-
-    const et = body.enriched_terms;
-
-    return {
-      term: et.Term,
-      fdr: et.FDR,
-      total: body.total_n_enriched_terms,
-      synonyms: et.Synonyms.map((list) => list?.split(";") || []),
-      coincident: et["Coincident Terms"].map((list) => list?.split(";") || []),
-      matchingGenes: et["Matching Genes in List"].map((geneList) => {
-        return geneList.split(" ");
-      }),
-    };
-  };
-
-  return fetchJson(
-    `/../../genetea-api/enriched-terms/?${query}`,
-    transformResponse as (o: object) => ReturnType<typeof transformResponse>
-  );
-}
-
-export async function fetchGeneTeaTermContext(
-  term: string,
-  genes: string[]
-): Promise<Record<string, string>> {
-  const query = qs.stringify(
-    { term, gene_list: genes, model: "v2", html: true },
-    { arrayFormat: "repeat" }
-  );
-
-  const transformResponse = (body: {
-    message?: string;
-    context: Record<string, string>;
-  }) => {
-    if (body.message) {
-      throw new Error(body.message);
-    }
-
-    return body.context;
-  };
-
-  return fetchJson(
-    `/../../genetea-api/context/?${query}`,
-    transformResponse as (o: object) => Record<string, string>
-  );
-}
-
 export async function fetchMetadataColumn(
   slice_id: string
 ): Promise<{
